when deleting waiter or manager check if exists first

+++ backend/django-server/authentication/signals.py
@@ -20,9 +20,9 @@ def signal_handler_when_user_is_added_or_removed_from_group(action, instance, pk
 
             if action == "pre_remove":
                 group = Group.objects.get(id=pk)
-                if group.name == Groups.WAITER.value:
+                if group.name == Groups.WAITER.value and Waiter.objects.filter(user=instance).exists():
                     waiter = Waiter.objects.get(user=instance)
                     waiter.delete()
-                elif group.name == Groups.MANAGER.value:
+                elif group.name == Groups.MANAGER.value and Manager.objects.filter(user=instance).exists():
                     manager = Manager.objects.get(user=instance)
                     manager.delete()
